+++ CHANGELOG.md
@@ -5,11 +5,13 @@ CHANGE LOG
 ## 4.2.0 (UPCOMING)
 
 * Support VPCs when creating droplets
+* Added missing firewall methods
 
 
 ## 4.1.1 (UPCOMING)
 
 * Fixed incorrect floating IP IDs
+* Fixed missing firewall client method
 
 
 ## 4.1.0 (25/01/2021)