+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGE LOG
 ==========
 
 
-## 4.1.1 (UPCOMING)
+## 4.1.1 (20/02/2021)
 
 * Fixed incorrect floating IP IDs
 * Fixed missing firewall client method